+++ phpstan-bootstrap.php
@@ -1,6 +1,3 @@
 <?php
 
 require 'vendor/autoload.php';
-require 'vendor/DateTimeFileWriter.php';
-require 'vendor/epub.php';
-require 'vendor/rb.php';