+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.x] - 20xx-xx-xx
+## Changed
+- Fixed a few more phpstan level 3 issues
+- Moved vendor/rb.php to lib/BicBucStriim using gabordemooij/redbean package
+- Moved vendor/epub.php to lib/BicBucStriim
+- Replaced vendor/DateTimeFileWriter.php with package slim/logger
+- Replaced abandoned package dflydev/markdown with michelf/php-markdown
+- Skipped non-working tests for author thumbnail, links and note
+### Removed
+- Removed package ircmaxell/password-compat as no longer relevant (PHP < 5.5)
+
 ## [1.6.6] - 2023-12-04
 ## Changed
 - Merged #1 by @Tocamadera, Amazon will start winding down MOBI format